@@ -1,8 +1,8 @@
 # This migration comes from attachinary (originally 20120612112526)
-class CreateAttachinaryTables < ActiveRecord::Migration
+class CreateAttachinaryTables < ActiveRecord::Migration[5.0]
   def change
     create_table :attachinary_files do |t|
-      t.references :attachinariable, polymorphic: true
+      t.references :attachinariable, polymorphic: true, index: false
       t.string :scope
 
       t.string :public_id
@@ -13,6 +13,6 @@ class CreateAttachinaryTables < ActiveRecord::Migration
       t.string :resource_type
       t.timestamps
     end
-    add_index :attachinary_files, [:attachinariable_type, :attachinariable_id, :scope], name: 'by_scoped_parent'
+    add_index :attachinary_files, [:attachinariable_type, :attachinariable_id, :scope], name: "by_scoped_parent"
   end
 end
