@@ -1,4 +1,4 @@
- class RenameTable < ActiveRecord::Migration
+ class RenameTable < ActiveRecord::Migration[5.0]
    def change
      rename_table :baskets, :basketlines
    end
