@@ -1,4 +1,4 @@
- class RenameTableBasket < ActiveRecord::Migration
+ class RenameTableBasket < ActiveRecord::Migration[5.0]
    def change
      rename_table :basketlines, :basket_lines
    end
