@@ -52,32 +52,34 @@ ActiveAdmin.register Order do
         @orders = Order.where(state: "paid").order(updated_at: :desc).page(params[:page]).per(20)
         @last_month_orders = Order.where("state LIKE ? AND updated_at >= ?", "paid", Time.now - 30 * 3600 * 24).count
         @last_six_month_orders = Order.where("state LIKE ? AND updated_at >= ?", "paid", Time.now - 180 * 3600 * 24).count
-        @basketlines_from_orders_lost = Order.where.not(state: "paid").map {|order| order.basketlines}.flatten
-        @categories_sales_data_six = categories_data(180, @basketlines_from_orders_lost)
-        @categories_sales_data_one = categories_data(30, @basketlines_from_orders_lost)
+        @categories_sales_data_six = categories_data(180)
+        @categories_sales_data_one = categories_data(30)
       end
     end
 
-    def categories_data(duration, basketlines_lost)
+    def categories_data(duration)
       categories_sales_data = {}
       Category.all.each do |category|
         sum = 0
-        categories_sales_data[category.id.to_s.to_sym] = {}
         category.ceramiques.each do |ceramique|
           ceramique.offer ? discount = ceramique.offer.discount : discount = 0
 
-          addition = ((ceramique.basketlines.where("updated_at >= ?", Time.now - duration * 3600 * 24) - basketlines_lost)
-          .map {|basketline| basketline.quantity}
-          .reduce(:+) || 0) * ceramique.price * (1 - discount)
+          addition = Basketline.where("basketlines.updated_at >= ? AND ceramique_id = ?", Time.now - duration * 3600 * 24, ceramique.id)
+            .left_outer_joins(:order).where("orders.state = ?", "paid").sum(:quantity) * ceramique.price * (1 - discount)
 
-          categories_sales_data[category.id.to_s.to_sym][ceramique.id.to_s.to_sym] = {}
-          categories_sales_data[category.id.to_s.to_sym][ceramique.id.to_s.to_sym][:"ceramique_sum"] = addition
-          categories_sales_data[category.id.to_s.to_sym][ceramique.id.to_s.to_sym][:"ceramique_name"] = ceramique.name
-          categories_sales_data[category.id.to_s.to_sym][ceramique.id.to_s.to_sym][:"ceramique_id"] = ceramique.id
+          if addition > 0
+            categories_sales_data[category.id.to_s.to_sym] = {}
+            categories_sales_data[category.id.to_s.to_sym][ceramique.id.to_s.to_sym] = {}
+            categories_sales_data[category.id.to_s.to_sym][ceramique.id.to_s.to_sym][:"ceramique_sum"] = addition
+            categories_sales_data[category.id.to_s.to_sym][ceramique.id.to_s.to_sym][:"ceramique_name"] = ceramique.name
+            categories_sales_data[category.id.to_s.to_sym][ceramique.id.to_s.to_sym][:"ceramique_id"] = ceramique.id
+          end
 
           sum += addition
         end
-        categories_sales_data[category.id.to_s.to_sym][:"sum"] = sum
+        if sum > 0
+          categories_sales_data[category.id.to_s.to_sym][:"sum"] = sum
+        end
       end
       categories_sales_data
     end
