@@ -23,7 +23,7 @@ $(document).ready(function() {
 
   function checkOffset(list_header_name, navbar_name) {
     var offset = $('.last-ceramique').offset().top - $(navbar_name).height() - 40
-    if ($('.univers-strip').size() > 0) {
+    if ($('.univers-strip').length > 0) {
       offset -=  180
     } else {
       offset +=  20
@@ -32,9 +32,9 @@ $(document).ready(function() {
   }
 
   $(document).scroll(function() {
-    if ($('#ceramique-filters-content').size() > 0) {
+    if ($('#ceramique-filters-content').length > 0) {
       changeButtonFixeNav('#ceramique-filters-content', '.navbar-laptop', '.list-header-1')
-    } else if ($('#ceramique-filters-content-darktheme').size() > 0) {
+    } else if ($('#ceramique-filters-content-darktheme').length > 0) {
       changeButtonFixeNav('#ceramique-filters-content-darktheme', '.navbar-laptop-darktheme', '.list-header-1-darktheme')
     }
   });
