@@ -1,3 +1,3 @@
-$(document).on('ready', function() {
+$(document).ready(function() {
   $('.stripe-button-el span').text("Payer par carte");
 });
