@@ -16,8 +16,8 @@ class OrdersController < ApplicationController
     if params[:quantity].to_i <= @ceramique.stock
       @basketline = Basketline.create(ceramique: @ceramique, quantity: params[:quantity].to_i, order: @order, with_support: params[:support] == "on")
       @ceramique.update(stock: @ceramique.stock - @basketline.quantity)
-      costs = Amountcalculation.new(@order).calculate_amount(@order)
-      @order.update(amount: costs[:total], port: costs[:port], ceramique: collect_ceramiques_for_stats)
+      costs = Amountcalculation.new(@order).calculate_amount(@order, current_user)
+      @order.update(amount: costs[:total], port: costs[:port], ceramique: collect_ceramiques_for_stats, weight: costs[:weight])
       flash[:notice] = "Votre panier sera conservé #{(ENV['BASKETDURATION'].to_f * 60).to_i } min"
       redirect_to order_path(@order)
     else
@@ -30,9 +30,13 @@ class OrdersController < ApplicationController
     if Order.find(params[:id])
       @order = Order.find(params[:id])
       if @order.state == "pending" || @order.state == "payment page"
-        @order.update(user: current_user) if current_user
+        if current_user
+          costs = Amountcalculation.new(@order).calculate_amount(@order, current_user)
+          @order.update(amount: costs[:total], port: costs[:port], weight: costs[:weight], user: current_user)
+        end
         @amount = @order.amount
         @port = @order.port
+        @weight = @order.weight
         render "show_#{@active_theme.name}"
       else
         flash[:notice] = "Votre panier a expiré"
@@ -50,8 +54,8 @@ class OrdersController < ApplicationController
     @ceramique = @basketline.ceramique
     @ceramique.update(stock: @ceramique.stock + @basketline.quantity)
     @basketline.destroy
-    costs = Amountcalculation.new(@order).calculate_amount(@order)
-    @order.update(amount: costs[:total], port: costs[:port], ceramique: @order.ceramique.sub(@ceramique.name+",",""))
+    costs = Amountcalculation.new(@order).calculate_amount(@order, current_user)
+    @order.update(amount: costs[:total], port: costs[:port], weight: costs[:weight], ceramique: @order.ceramique.sub(@ceramique.name+",",""))
     if @order.basketlines.present?
       redirect_to order_path(@order)
     else
@@ -64,7 +68,8 @@ class OrdersController < ApplicationController
   private
 
   def create_order
-    order  = Order.create!(ceramique: @ceramique.name, state: 'pending', take_away: false)
+    order  = Order.new(ceramique: @ceramique.name, state: 'pending', take_away: false)
+    order.save
     session[:order] = order.id
     return order
   end
