@@ -2,8 +2,8 @@ class Order < ApplicationRecord
   monetize :amount_cents
   monetize :port_cents
   has_many :basketlines
-  belongs_to :user
-  belongs_to :lesson
+  belongs_to :user, required: false
+  belongs_to :lesson, required: false
 
   validates_inclusion_of :take_away, in: [true, false]
 end
