@@ -5,7 +5,7 @@ class Amountcalculation
     @order
   end
 
-  def calculate_amount(order)
+  def calculate_amount(order, user)
     amount_ceramique = 0
     total_weight = 0
     order.basketlines.each do |basketline|
@@ -14,15 +14,30 @@ class Amountcalculation
       amount_ceramique += (basketline.ceramique.price * (1 - ceramique_discount)) * basketline.quantity + amount_support * basketline.quantity
       total_weight += basketline.ceramique.weight * basketline.quantity
     end
-    if total_weight > 0
-      tarif_colis = HTTParty.get(
-        "https://api.laposte.fr/tarifenvoi/v1?type=colis&poids=#{total_weight}",
-        headers: {"X-Okapi-Key" => ENV['LAPOSTE_API_KEY'] }
-      )
-      return {total: amount_ceramique, port: tarif_colis[0]["price"].to_money}
+    if user
+      if user.country == "FR" && total_weight > 0
+        fr_shipping_cost(amount_ceramique, total_weight)
+      elsif total_weight > 0
+        shipping_cost = ShippingCategory.where(alpha2: user.country).where("weight >= ?", total_weight).min.price_cents.to_f / 100
+        {total: amount_ceramique, port: shipping_cost.to_money, weight: total_weight}
+      else
+        {total: 0, port: 0, weight: 0}
+      end
     else
-      {total: 0, port: 0}
+      if total_weight > 0
+        fr_shipping_cost(amount_ceramique, total_weight)
+      else
+        {total: 0, port: 0, weight: 0}
+      end
     end
+  end
+
+  def fr_shipping_cost(amount_ceramique, total_weight)
+    tarif_colis = HTTParty.get(
+      "https://api.laposte.fr/tarifenvoi/v1?type=colis&poids=#{total_weight}",
+      headers: {"X-Okapi-Key" => ENV['LAPOSTE_API_KEY'] }
+    )
+    return {total: amount_ceramique, port: tarif_colis[0]["price"].to_money, weight: total_weight}
   end
 
 end
