@@ -1,6 +1,6 @@
 require_relative 'boot'
 
-require "rails"
+require 'rails/all'
 # Pick the frameworks you want:
 require "active_model/railtie"
 require "active_job/railtie"
@@ -11,7 +11,6 @@ require "action_view/railtie"
 require "action_cable/engine"
 require "sprockets/railtie"
 require "attachinary/orm/active_record"
-# require "rails/test_unit/railtie"
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -23,13 +22,16 @@ module Ceramiquesnugier
       generate.assets false
       generate.helper false
     end
+    # Initialize configuration defaults for originally generated Rails version.
+    config.load_defaults 5.0
 
     config.exceptions_app = self.routes
     config.action_view.embed_authenticity_token_in_remote_forms = true
     config.i18n.load_path += Dir[Rails.root.join('locales', '*.{rb,yml}').to_s]
     config.i18n.default_locale = :fr
     # Settings in config/environments/* take precedence over those specified here.
-    # Application configuration should go into files in config/initializers
-    # -- all .rb files in that directory are automatically loaded.
+    # Application configuration can go into files in config/initializers
+    # -- all .rb files in that directory are automatically loaded after loading
+    # the framework and any gems in your application.
   end
 end
