@@ -1,6 +1,6 @@
-%w(
+%w[
   .ruby-version
   .rbenv-vars
   tmp/restart.txt
   tmp/caching-dev.txt
-  config/application.yml\n).each { |path| Spring.watch(path) }
+].each { |path| Spring.watch(path) }
